winter_air

Module proof

Contains STARK proof struct and associated components.

Structs

Commitments

Commitments made by the prover during commit phase of the protocol.

Context

Basic metadata about a specific execution of a computation.

OodFrame

Trace and constraint polynomial evaluations at an out-of-domain point.

Proof

A proof generated by Winterfell prover.

Queries

Decommitments to evaluations of a set of functions at multiple points.

Table

A two-dimensional table of field elements arranged in row-major order.

TraceOodFrame

Trace evaluation frame at the out-of-domain point.

Functions

ceil

log2

powf

sqrt