winter_air

Trait GkrVerifier

pub trait GkrVerifier {
    type GkrProof: Deserializable;
    type Error: ToString;

    // Required method
    fn verify<E, Hasher>(
        &self,
        gkr_proof: Self::GkrProof,
        public_coin: &mut impl RandomCoin<BaseField = E::BaseField, Hasher = Hasher>,
    ) -> Result<GkrRandElements<E>, Self::Error>
       where E: FieldElement,
             Hasher: ElementHasher<BaseField = E::BaseField>;
}

A trait for verifying a GKR proof.

Specifically, the use case in mind is proving the constraints of a LogUp bus using GKR, as described in Improving logarithmic derivative lookups using GKR.

Required Associated Types

type GkrProof: Deserializable

The GKR proof.

type Error: ToString

The error that can occur during GKR proof verification.

Required Methods

fn verify<E, Hasher>( &self, gkr_proof: Self::GkrProof, public_coin: &mut impl RandomCoin<BaseField = E::BaseField, Hasher = Hasher>, ) -> Result<GkrRandElements<E>, Self::Error>
where E: FieldElement, Hasher: ElementHasher<BaseField = E::BaseField>,

Verifies the GKR proof, and returns the random elements that were used in building the Lagrange kernel auxiliary column.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.